+++ pkgs/ffigen/CHANGELOG.md
@@ -3,9 +3,10 @@
 - __Breaking change__: Code-gen the ObjC `id` type to `ObjCObjectBase` rather
   than `NSObject`, since not all ObjC classes inherit from `NSObject`. Eg
   `NSProxy`.
+- Rename ObjC interface methods that clash with type names. Fixes
+  https://github.com/dart-lang/native/issues/1007.
 - __Breaking change__: Update Enum types to use the integer type returned by
   libclang instead of `Int32`.
-
 
 ## 12.0.0